Having been unable to figure out how to create a `PeopleDocument` during testing I have rebuilt the test to validate the logic at the FileRep level. I have kept the old tests and fixtures commented out for now.
